+++ core/CHANGELOG.md
@@ -7,6 +7,7 @@ Unreleased
 _YYYY-MM-DD_
 
 * Dependency loop error now lists all test cases that formed a cycle
+* ConsoleFormat (color, intensity) can be customized via environment (#298)
 
 Version 1.4.3
 ---------------